Moves files into include and src
